requirements.txt
numpy==1.23.1
pandas==1.5.0
Pillow==9.3.0
timm==0.6.11
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
